# Conflicts:
#	pom.xml
#	src/main/java/io/khasang/moika/config/AppConfig.java
#	src/main/java/io/khasang/moika/controller/AppController.java
#	src/main/java/io/khasang/moika/dao/impl/CompanyDaoImpl.java

+++ src/main/java/io/khasang/moika/service/CompanyService.java
@@ -9,7 +9,7 @@ public interface CompanyService {
 
     void addCompany(Company company, String name);
 
-    Company getCompanyById(int id);
+    Company getCompanyById(long id);
 
     List<Company> getCompanyGazpromList();
 }
